Adapter
Convert USD to EUR using Adapter pattern

The commit modifies the app by implementing a currency converter that allows the conversion of values from US dollars (USD) to euros (EUR). Previously, the app only supported the conversion from USD to British pounds (GBP). The new functionality is achieved by implementing the Adapter pattern. The class `OldCurrencyConverter` is an old converter that supports the conversion from USD to GBP. The class `CurrencyAdapter` acts as an intermediary and adapts the interface of `OldCurrencyConverter` to provide direct conversion from USD to EUR. The conversion rates used are as follows: 1 USD = 0.85 EUR and 1 USD = 0.80 GBP. The commit also updates the main function to demonstrate the usage of the adapter to convert a given amount in USD to EUR.